LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/parsing/column-rule-width-computed-expected.txt


PASS Property column-rule-width value 'calc(10px + 0.5em)'
PASS Property column-rule-width value 'calc(10px - 0.5em)'
PASS column-rule-width is as specified when column-rule-style is none or hidden
